[REFACTOR/#1552] `throttledNoRippleClickable` 공통 모듈로 이동
- `ThrottledNoRippleClickable.kt` 파일을 `feature/soptamp`에서 `core/common` 모듈로 이전하여 재사용성 확보
- `Toolbar`, `UserMissionListScreen`, `RankingScreen` 등 기존 참조 위치를 새로운 패키지 경로(`org.sopt.official.common.util`)로 업데이트
